third_party/blink/web_tests/external/wpt/shadow-dom/declarative/declarative-shadow-dom-slot-assignment-serialization-expected.txt

This is a testharness.js-based test.
[FAIL] shadowrootslotassignment=manual is serialized and appears before shadowrootclonable and shadowrootserializable
  assert_equals: expected "manual" but got "named"
[FAIL] shadowrootslotassignment=manual serializes between shadowrootdelegatesfocus and shadowrootclonable
  assert_equals: expected "manual" but got "named"
Harness: the test ran to completion.
